update k cho xóa sản phẩm khi đã được mua

+++ src/main/java/com/fpoly/api/ProductApi.java
@@ -4,7 +4,9 @@ import com.fpoly.dto.request.ProductRequest;
 import com.fpoly.dto.request.ProductSearch;
 import com.fpoly.dto.response.ProductResponse;
 import com.fpoly.entity.Product;
+import com.fpoly.exception.MessageException;
 import com.fpoly.mapper.ProductMapper;
+import com.fpoly.repository.InvoiceDetailRepository;
 import com.fpoly.servive.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -30,6 +32,10 @@ public class ProductApi {
     @Autowired
     private ProductMapper productMapper;
 
+    @Autowired
+    private InvoiceDetailRepository invoiceDetailRepository;
+
+
     @PostMapping("/admin/create")
     public ResponseEntity<?> save(@Valid @RequestBody ProductRequest productRequest, BindingResult bindingResult){
         if(bindingResult.hasErrors()){
@@ -51,6 +57,12 @@ public class ProductApi {
 
     @DeleteMapping("/admin/delete")
     public ResponseEntity<?> delete(@RequestParam("id") Long id){
+        if (invoiceDetailRepository.countByProduct(id) > 0 ||
+                invoiceDetailRepository.countByProductColor(id) > 0 ||
+                invoiceDetailRepository.countByProductSize(id) > 0) {
+            throw new MessageException("Sản phẩm đã có người mua, không thể xóa");
+        }
+
         productService.delete(id);
         return new ResponseEntity<>("delete success", HttpStatus.OK);
     }
